restored url hash for moving active camera to specific lat/long coordinates
