+++ 3DAmsterdam/Assets/Amsterdam3D/Scripts/Camera/CameraModeChanger.cs
@@ -1,5 +1,6 @@
 ﻿
 using BruTile.Wms;
+using ConvertCoordinates;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -75,6 +76,16 @@ namespace Amsterdam3D.CameraMotion
             {
                 StartCoroutine(streetView.MoveToPosition(cameraTransform, position, rotation));
             }
+        }
+
+        public void ChangedPointFromUrl(string latLong)
+        {
+            string[] coordinates = latLong.Split(',');
+            var lat = double.Parse(coordinates[0]);
+            var lon = double.Parse(coordinates[1]);
+
+            var convertedCoordinate = CoordConvert.WGS84toUnity(lon, lat);
+            currentCamera.transform.position = new Vector3(convertedCoordinate.x, this.transform.position.y, convertedCoordinate.z);
         }
 
         public void GodViewMode()
