+++ 3DAmsterdam/Assets/Amsterdam3D/Scripts/Camera/GodViewCamera.cs
@@ -130,16 +130,6 @@ namespace Amsterdam3D.CameraMotion
             focusingOnTargetPoint(targetLocation);
         }
 
-        public void ChangedPointFromUrl(string latLong)
-        {
-            string[] coordinates = latLong.Split(',');
-            var lat = double.Parse(coordinates[0]);
-            var lon = double.Parse(coordinates[1]);
-
-            var convertedCoordinate = CoordConvert.WGS84toUnity(lon, lat);
-            this.transform.position = new Vector3(convertedCoordinate.x, this.transform.position.y, convertedCoordinate.z);
-        }
-
         void HandleTranslationInput()
         {         
             moveSpeed = Mathf.Sqrt(camera.transform.position.y) * speedFactor;
