+++ 3DAmsterdam/Assets/WebGLTemplates/Responsive3DAmsterdam/index.html
@@ -91,7 +91,7 @@
 	  }
 
 	  function locationHashChanged() {
-		  unityInstance.SendMessage("MainCamera", "ChangedPointFromUrl", window.location.hash.replace("#", ""));
+		  unityInstance.SendMessage("CameraModeChanger", "ChangedPointFromUrl", window.location.hash.replace("#", ""));
 	  }
 
 	  function SelectSharedURL() {
